digital_assets/Swarm_Finance_Agent/requirements.txt
agno>=2.2.10
yfinance
duckduckgo-search
python-dotenv
